[Homework] Script 3 additional Classes (1/3) [PlayerHealth]
Based on what you've learned so far in the class, create three new Scripts with the following class names:
	•PlayerHealth
	•PlayerScore
	•CharacterMover
I created the PlayerHealth C# script.
I chose to create the playerHealth variable as public so it could be seen in the unity editor.
I  also initialized it at 100.
I added the script as a component to the main player so now the playerHealth variable can be seen when testing the game.