[Homework] Script 3 additional Classes (2/3) [PlayerScore]
Based on what you've learned so far in the class, create three new Scripts with the following class names:
	•PlayerHealth
	•PlayerScore
	•CharacterMover
- I created the PlayerScore C# script.
I want two scores for this game. 1) Coins collected for the run, 2) Distance Ran.
I created the following variables:
	- int coinCounter
	- float distanceTraveled
	- float gameStartingTime
	- float distanceScaler
Using these I made it so the distanceTraveled goes up over time. I know it is not accurate on the actual meters traveled, but I think it goes up at a good scale for the pace.
I also started a coinCounter to keep track of and hopefully increment the coins when I hit them.